title/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-jazzmin==3.0.3
djangorestframework==3.16.1
PyMySQL==1.1.2
sqlparse==0.5.5
tzdata==2025.3
